requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1

